% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magic_balanced.R
\name{magic.balanced}
\alias{magic.balanced}
\title{Check if a MAGIC design is balanced.}
\usage{
magic.balanced(fmat, verbose = FALSE)
}
\arguments{
\item{fmat}{a matrix of founder combinations.}

\item{verbose}{a logical indicator of whether verbose output is desired.}
}
\value{
a logical indicator of whether the MAGIC design is balanced or not, and
also a table of all founder pairings if \code{verbose=TRUE}.
}
\description{
This function takes a matrix of founder combinations and determines if the design
is balanced or not. The matrix rows are funnels and columns are founders.
}
\examples{
\donttest{
mpop <- magic.partial(n=8, m=1, balanced=TRUE)
magic.balanced(mpop$fcomb[[3]])
}

}
