% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magic_deficient.R
\name{magic.deficient}
\alias{magic.deficient}
\title{Create a deficient MAGIC design.}
\usage{
magic.deficient(n, m, balanced, inbred = T)
}
\arguments{
\item{n}{number of founders.}

\item{m}{number of funnel sets (balanced=T) or funnels (balanced=F).}

\item{balanced}{logical indicator of whether balanced partial design is chosen or ignored.}

\item{inbred}{logical indicator of whether the founders are inbred (default to TRUE).}
}
\value{
an object of "cross.info" class, which is a list of
founder combinations (fcomb) and crossing plans (xplan).
}
\description{
This function produces m (or m-set) founder combinations and crosses for
number of founders that is not a power of 2 (e.g. 3,5,6,7,9,...).
}
\examples{
\donttest{
mpop <- magic.deficient(n=5, m=2, balanced=T)
}

}
