% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magic_full.R
\name{magic.full}
\alias{magic.full}
\title{Create a full MAGIC design.}
\usage{
magic.full(n, inbred = TRUE)
}
\arguments{
\item{n}{number of founders (4, 8).}

\item{inbred}{logical indicator of whether the founders are inbred.}
}
\value{
an object of "cross.info" class, \emph{i.e.} a list of
founder combinations (fcomb) and crossing plans (xplan).
}
\description{
This function produces all possible founder combinations and crosses for
only either 4 or 8 founders. In a full design, the number of unique
founder combinations (funnels) is given by \eqn{n!/2^{n-1}}. Full design for
16 or more founders are impossible in practice, for example, a full design
for 16 founders would have 638,512,875 funnels.
}
\examples{
\donttest{
mpop <- magic.full(n=4)
}

}
