% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magic_partial.R
\name{magic.partial}
\alias{magic.partial}
\title{Create a partial MAGIC design.}
\usage{
magic.partial(n, m, balanced, n.try = 1000, inbred = T)
}
\arguments{
\item{n}{number of founders.}

\item{m}{number of funnel sets (balanced=T) or funnels (balanced=F).}

\item{balanced}{logical indicator of whether balanced partial design is chosen or ignored.}

\item{n.try}{number of attempts to find balanced partial design (default to 1000, applicable only if n=8 and balanced=T).}

\item{inbred}{logical indicator of whether the founders are inbred (default to TRUE).}
}
\value{
an object of "cross.info" class, which is a list of
founder combinations (fcomb) and crossing plans (xplan).
}
\description{
This function produces m (or m-set) founder combinations and crosses for
8 or more founders.
}
\examples{
\donttest{
mpop <- magic.partial(n=8, m=1, balanced=T)
}

}
