% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magic_ped.R
\name{magic.ped}
\alias{magic.ped}
\title{Convert a "cross.info" object into a pedigree.}
\usage{
magic.ped(xinfo)
}
\arguments{
\item{xinfo}{an object of "cross.info" class.}
}
\value{
a pedigree.
}
\description{
This function takes the "cross.info" object and converts it into a pedigree that
has 4 columns: individual ID, parent 1 ID, parent 2 ID and generation number.
}
\examples{
\donttest{
mpop <- magic.partial(n=8, m=1, balanced=T)
mped <- magic.ped(xinfo=mpop)
}

}
