% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magic_ped2cross.R
\name{magic.ped2cross}
\alias{magic.ped2cross}
\title{Convert a pedigree into a crossing plan.}
\usage{
magic.ped2cross(ped)
}
\arguments{
\item{ped}{a pedigree.}
}
\value{
a list of crossing plans for each crossing generation (xplan).
}
\description{
This function takes a user provided pedigree and converts it into a crossing plan
that can be used in the simulation. The pedigree must have four columns in the
following order: individual ID, parent 1 ID, parent 2 ID and generation number.
}
\examples{
\donttest{
mped <- cbind(ind=1:13,
              p1=c(0,0,0,0,1,1,1,2,2,3,5,6,7),
              p2=c(0,0,0,0,2,3,4,3,4,4,10,9,8),
              gen=c(0,0,0,0,1,1,1,1,1,1,2,2,2))
mpop <- magic.ped2cross(ped=mped)
}

}
