% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magic_ped2plot.R
\name{magic.ped2plot}
\alias{magic.ped2plot}
\title{Plot pedigree.}
\usage{
magic.ped2plot(
  ped,
  basic = F,
  show.partial = F,
  w2h.ratio = 2,
  force.option = F
)
}
\arguments{
\item{ped}{a pedigree.}

\item{basic}{a logical indicator of whether it is a basic design or not.}

\item{show.partial}{a logical indicator of whether to plot a partial pedigree or not.}

\item{w2h.ratio}{a numerical value of width-to-height-ratio in the pedigree plot.}

\item{force.option}{a logical indicator of whether to force the function to plot
using a simpler option.}
}
\value{
a HTML file of an interactive pedigree plot.
}
\description{
This function plots the pedigree based on the type of design.
By default, the function will plot a comprehensive pedigree for
\code{n} of 4, 8, 16, 32, 64, 128 and a simple pedigree for other \code{n}.
The pedigree is generated from either \link{magic.ped} or \link{magic.eval}.
Note: if the MAGIC design is created from a user-provided pedigree,
then the function will always plot a simple pedigree.
}
\examples{
\donttest{
mpop <- magic.eval(n=8, m=1, reps=c(1,1,2), self=c(0,0,3), balanced=T)
magic.ped2plot(ped=mpop$ped)
}

}
