% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magic_plot.R
\name{magic.plot}
\alias{magic.plot}
\title{Plot and compare different MAGIC designs.}
\usage{
magic.plot(
  input,
  display = c("interval", "whole"),
  fpair = NULL,
  chr.names = NULL
)
}
\arguments{
\item{input}{a list of outputs from magic.eval().}

\item{display}{a character indicator of whether "interval" or "whole" plot display is desired.}

\item{fpair}{a matrix of two columns of founder pairs (ignored if display="whole").}

\item{chr.names}{a vector of chromosome names (optional).}
}
\value{
multi-section plots.
}
\description{
This function takes a list of outputs from magic.eval() and compares the MAGIC designs
via various plots. There are two plot display option available: "interval" highlights
the recombinations within the previously specified hap.int argument, while "whole"
shows the distribution of founder genomes in the RILs.
}
\examples{
\donttest{
mpop1 <- magic.eval(n=8, m=1, reps=c(1,1,4), self=c(0,0,3), balanced=T, n.sim=10)
mpop2 <- magic.eval(n=8, m=7, reps=c(1,1,4), self=c(0,0,3), balanced=F, n.sim=10)
magic.plot(input=list(mpop1, mpop2), display="interval")
}

}
\seealso{
\link{magic.summary}
}
