% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magic_rearrange.R
\name{magic.rearrange}
\alias{magic.rearrange}
\title{Rearrange the founder combinations.}
\usage{
magic.rearrange(fmat)
}
\arguments{
\item{fmat}{matrix of founder combinations.}
}
\value{
an object of rearranged matrix of founder combinations.
}
\description{
This function rearranges a matrix of founder combinations such that smaller
founder appears before larger founder in any cross (e.g. 3124 becomes 1324).
}
\examples{
\donttest{
mpop <- t(sapply(1:3, FUN=function(x) sample(1:8, 8)))
mpop <- magic.rearrange(fmat=mpop)
}

}
