% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magic_reps.R
\name{magic.reps}
\alias{magic.reps}
\title{Add replicates to crosses.}
\usage{
magic.reps(xinfo, reps)
}
\arguments{
\item{xinfo}{an object of "cross.info" class.}

\item{rep}{a vector of replicates in each crossing generation.}
}
\value{
an object of "cross.info" class, which is a list of
founder combinations (fcomb) and crossing plans (xplan).
}
\description{
This function takes the "cross.info" object and add replicates to the crosses.
}
\examples{
\donttest{
mpop <- magic.partial(n=8, m=1, balanced=T)
mpop <- magic.reps(xinfo=mpop, reps=c(1,1,2))
}

}
