% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magic_self.R
\name{magic.self}
\alias{magic.self}
\title{Add selfing after crossing.}
\usage{
magic.self(xinfo, self)
}
\arguments{
\item{xinfo}{an object of "cross.info" class.}

\item{self}{a vector of number of selfing generations after each crossing generation.}
}
\value{
an object of "cross.info" class, \emph{i.e.} a list of
founder combinations (fcomb) and crossing plans (xplan).
}
\description{
This function takes the "cross.info" object and add selfing after crosses. In the
usual design, the only selfing is the single-seed descent after the last crossing
generation.
}
\examples{
\donttest{
mpop <- magic.partial(n=8, m=1, balanced=T)
mpop <- magic.self(xinfo=mpop, self=c(0,0,3))
}

}
