% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magic_sim.R
\name{magic.sim}
\alias{magic.sim}
\title{Simulate populations based on the MAGIC design.}
\usage{
magic.sim(
  xplan,
  inbred = T,
  marker.dist = 0.01,
  chr.len = c(1, 2),
  n.sim = 1,
  hap.int = 0.05,
  n.hap = 1,
  keep = F
)
}
\arguments{
\item{xplan}{a list of crossing plans (xplan).}

\item{inbred}{a logical indicator of whether the founders are inbred (default is TRUE).}

\item{marker.dist}{a numerical value of marker distance in Morgan (default is 0.01).}

\item{chr.len}{a vector of chromosome lengths in Morgan (default is c(1,2)).}

\item{n.sim}{an integer of number of simulations (default is 1).}

\item{hap.int}{a numerical value of marker interval for evaluating haplotypes (default is 0.05)).}

\item{n.hap}{an integer of 1 or 2 haploid marker data of each RIL are used (default is 1).}

\item{keep}{a logical indicator of whether to export the marker data (default is F).}
}
\value{
a list of simulation summary.
}
\description{
This function takes the crossing plans (xplan) and simulate populations based on
the crossing plans. Simulation is done using AlphaSimR. The founder genotypes in
the RILs from each simulation are summarized. Simulated marker data can also be
exported, but preferably with n.sim set to 1 or any small value as the output file
can be large.
}
\examples{
\donttest{
mpop <- magic.partial(n=8, m=1, balanced=T)
mpop <- magic.sim(xplan=mpop$xplan)
}

}
