% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magic_summary.R
\name{magic.summary}
\alias{magic.summary}
\title{Tabulate the comparisons of different MAGIC designs.}
\usage{
magic.summary(input)
}
\arguments{
\item{input}{a list of outputs from \link{magic.eval}.}
}
\value{
a list of 6 different comparisons.
}
\description{
This function takes a list of outputs from \link{magic.eval} and compares the MAGIC designs.
It is similar to \link{magic.plot}, excepts the results are shown in tables instead of figures.
}
\examples{
\donttest{
mpop1 <- magic.eval(n=8, m=1, reps=c(1,1,4), self=c(0,0,3), balanced=T, n.sim=10)
mpop2 <- magic.eval(n=8, m=7, reps=c(1,1,4), self=c(0,0,3), balanced=F, n.sim=10)
mpop <- magic.summary(input=list(mpop1, mpop2))
}

}
\seealso{
\link{magic.plot}
}
