% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magicdesignee.R
\name{magicdesignee}
\alias{magicdesignee}
\title{Launch magicdesignee.}
\usage{
magicdesignee()
}
\description{
This function launches a Shiny app (magicdesignee) locally. The app implements
magicdesign in its backend.
}
