% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magic_eval.R
\name{magic.eval}
\alias{magic.eval}
\title{Main function to design and test MAGIC population.}
\usage{
magic.eval(
  ped = NULL,
  n = NULL,
  m = NULL,
  reps = NULL,
  self = NULL,
  inbred = T,
  balanced = F,
  minimize = F,
  n.try = 1000,
  addx = NULL,
  repx = 1,
  selfx = 3,
  marker.dist = 0.01,
  chr.len = c(1, 2),
  n.sim = 1,
  hap.int = 0.05,
  n.hap = 1,
  keep = F
)
}
\arguments{
\item{ped}{a pedigree with 4 columns: individual ID, parent 1 ID, parent 2 ID, generation
number, in the format of either matrix or data.frame (default is NULL).}

\item{n}{an integer of number of founders (default is NULL).}

\item{m}{an integer of number of funnel sets if balanced=T or number of funnels if
balanced=F (default is NULL).}

\item{reps}{a vector of replicates in each crossing generation (default is NULL).}

\item{self}{a vector of number of generations to self after crossing (default is NULL).}

\item{inbred}{a logical indicator of whether the founders are inbred (default is TRUE).}

\item{balanced}{a logical indicator of whether a balanced partial design is desired
(default is FALSE).}

\item{minimize}{a logical indicator of whether to minimize crossing (default is NULL).}

\item{n.try}{an integer of number of attempts to find balanced partial design (default
is 1000 and ignored if balanced=F or n > 8).}

\item{addx}{an integer of either 1 or 2 indicating the type of additional crosses
(default is NULL).}

\item{repx}{an integer of number of replicates in the additional crossing (default is 1).}

\item{selfx}{an integer of number of generations to self after additional crossing
(default is 3).}

\item{marker.dist}{a numerical value of marker distance in Morgan (default is 0.01).}

\item{chr.len}{a vector of chromosome lengths in Morgan (default is c(1,2)).}

\item{n.sim}{an integer of number of simulations (default is 1).}

\item{hap.int}{a numerical value of marker interval for evaluating haplotypes (default
is 0.05).}

\item{n.hap}{an integer of 1 or 2 haploid marker data of each RIL are used (default is 1).}

\item{keep}{a logical indicator of whether to export the marker data (default is F).}
}
\value{
a list of simulation summary.
}
\description{
This function takes various input arguments from user and simulate the desired
MAGIC population. The MAGIC population can be designed from using either a pedigree
or user-defined arguments, in which the latter includes MAGIC population types like
full, partial, basic or deficient, as well as designs that are balanced or unbalanced.
}
\examples{
\donttest{
mpop <- magic.eval(n=8, m=1, reps=c(1,1,2), self=c(0,0,3), balanced=T)
}

}
