% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magic_full.R
\name{magic.full}
\alias{magic.full}
\title{Create a full MAGIC design.}
\usage{
magic.full(n, inbred = T)
}
\arguments{
\item{n}{number of founders.}

\item{inbred}{logical indicator of whether the founders are inbred (default to TRUE).}
}
\value{
an object of "cross.info" class, which is a list of
founder combinations (fcomb) and crossing plans (xplan).
}
\description{
This function produces all possible founder combinations and crosses for
only either 4 or 8 founders. In a full design, the number of unique
founder combinations (funnels) is given by n!/(2^(n-1)). Full design for
16 or more founders are impossible in practice, as 16 founders would
result in 638,512,875 funnels.
}
\examples{
\donttest{
mpop <- magic.full(n=4)
}

}
