% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magic_ped2plot.R
\name{magic.ped2plot}
\alias{magic.ped2plot}
\title{Plot pedigree generated from magic.eval().}
\usage{
magic.ped2plot(
  ped,
  filename,
  basic = F,
  show.partial = F,
  w2h.ratio = 2,
  force.option = F
)
}
\arguments{
\item{ped}{a pedigree generated from magic.eval().}

\item{filename}{a string of filename to save the pedigree plot file.}

\item{basic}{a logical indicator of whether it is a basic design or not.}

\item{show.partial}{a logical indicator of whether to plot a partial pedigree or not.}

\item{w2h.ratio}{a numerical value of width-to-height-ratio in the pedigree plot
(default is 2).}

\item{force.option}{a logical indicator of whether to force the function to plot
using a simpler option (default is FALSE).}
}
\value{
a HTML file of an interactive pedigree plot.
}
\description{
This function plots the pedigree based on the type of design.
}
\examples{
\donttest{
mpop <- magic.eval(n=8, m=1, reps=c(1,1,2), self=c(0,0,3), balanced=T)
magic.ped2plot(ped=mpop$ped, filename="filename")
}

}
