% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magic_plot.R
\name{magic.plot}
\alias{magic.plot}
\title{Plot and compare different MAGIC designs.}
\usage{
magic.plot(
  input,
  display = c("interval", "whole"),
  fpair = NULL,
  chr.names = NULL,
  annotate = T,
  design.names = NULL
)
}
\arguments{
\item{input}{a list of one or more outputs from \link{magic.eval}.}

\item{display}{a character indicator of whether \code{"interval"} or \code{"whole"} plot display is desired.}

\item{fpair}{a matrix of two columns of founder pairs (ignored if \code{display="whole"}).}

\item{chr.names}{a vector of chromosome names (optional).}

\item{annotate}{a logical indicator of whether to annotate subplot C if \code{display="whole"}.}

\item{design.names}{a vector of design names (optional).}
}
\value{
multi-section plots.
}
\description{
This function takes a list of outputs from \link{magic.eval} and compares the MAGIC designs
via various plots. There are two plot display options available: \code{"interval"} highlights
the recombinations within the previously specified \code{hap.int} argument, while \code{"whole"}
shows the distribution of founder genomes in the RILs.
}
\examples{
\donttest{
mpop1 <- magic.eval(n=8, m=1, reps=c(1,1,4), self=c(0,0,3), balanced=T, n.sim=10)
mpop2 <- magic.eval(n=8, m=7, reps=c(1,1,4), self=c(0,0,3), balanced=F, n.sim=10)
magic.plot(input=list(mpop1, mpop2), display="interval")
}

}
\seealso{
\link{magic.summary}
}
