% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magic_sim.R
\name{magic.sim}
\alias{magic.sim}
\title{Simulate populations based on the MAGIC design.}
\usage{
magic.sim(
  xplan,
  inbred = T,
  marker.dist = 0.01,
  chr.len = c(1, 2),
  n.sim = 1,
  hap.int = 0.05,
  n.hap = 1,
  keep = F
)
}
\arguments{
\item{xplan}{a list of crossing plans (xplan).}

\item{inbred}{a logical indicator of whether the founders are inbred.}

\item{marker.dist}{a numerical value of marker distance in Morgan.}

\item{chr.len}{a vector of chromosome lengths in Morgan.}

\item{n.sim}{an integer of number of simulations.}

\item{hap.int}{a numerical value of marker interval for evaluating haplotypes.}

\item{n.hap}{an integer of 1 or 2 haploid marker data of each RIL are used.}

\item{keep}{a logical indicator of whether to keep the marker data.}
}
\value{
a list of simulation summary.
}
\description{
This function takes the crossing plans (xplan) and simulates populations based on
the crossing plans. Simulation is done using
\href{https://cran.r-project.org/web/packages/AlphaSimR/index.html}{AlphaSimR}. The founder genotypes in
the RILs from each simulation are summarized. By default, the simulated marker data is
not kept. If the simulated marker data is really required, please set
\code{n.sim} to 1 or any small value as the output object can be large.
}
\examples{
\donttest{
mpop <- magic.partial(n=8, m=1, balanced=T)
mpop <- magic.sim(xplan=mpop$xplan)
}

}
